+++ common/src/main/java/io/github/flemmli97/runecraftory/common/blocks/tile/CropBlockEntity.java
@@ -2,17 +2,21 @@ package io.github.flemmli97.runecraftory.common.blocks.tile;
 
 import io.github.flemmli97.runecraftory.api.datapack.CropProperties;
 import io.github.flemmli97.runecraftory.common.blocks.BlockCrop;
+import io.github.flemmli97.runecraftory.common.config.GeneralConfig;
 import io.github.flemmli97.runecraftory.common.registry.ModBlocks;
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.protocol.Packet;
 import net.minecraft.network.protocol.game.ClientGamePacketListener;
 import net.minecraft.network.protocol.game.ClientboundBlockEntityDataPacket;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.util.Mth;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
+
+import java.util.Random;
 
 public class CropBlockEntity extends BlockEntity {
 
@@ -20,7 +24,7 @@ public class CropBlockEntity extends BlockEntity {
     private float age;
     private long lastGrowth = 0;
     private float cropLvl;
-    private boolean withered;
+    private boolean wilted;
     private BlockPos[] nearbyGiant;
 
     public CropBlockEntity(BlockPos blockPos, BlockState blockState) {
@@ -57,26 +61,53 @@ public class CropBlockEntity extends BlockEntity {
     }
 
     public void growCrop(Level level, BlockPos pos, BlockState state, float speed, float cropLvl, float seasonModifier) {
+        if (this.wilted)
+            this.wilted = false;
         Block block = state.getBlock();
         this.age += speed * seasonModifier;
-        BlockState newState = state;
         if (block instanceof BlockCrop crop) {
             float max = crop.properties().map(CropProperties::growth).orElse(0);
             this.age = Math.min(crop.properties().map(CropProperties::growth).orElse(0), this.age);
             int stateAge = (int) (this.age * 3 / max);
-            newState = state.setValue(BlockCrop.AGE, Mth.clamp(stateAge, 0, 3));
+            state = state.setValue(BlockCrop.AGE, Mth.clamp(stateAge, 0, 3)).setValue(BlockCrop.WILTED, this.wilted);
+            level.setBlock(pos, state, Block.UPDATE_ALL);
         }
-        level.sendBlockUpdated(pos, state, newState, Block.UPDATE_ALL);
         this.setChanged();
     }
 
-    public void setWithered() {
-        this.withered = true;
-        //this.world.setBlockState(this.pos, this.world.getBlockState(this.pos).withMirror(mirrorIn))
+    public void tryWitherCrop(Random rand) {
+        if (rand.nextFloat() < GeneralConfig.witherChance) {
+            if (this.isWilted()) {
+                if (this.level != null)
+                    this.level.setBlock(this.getBlockPos(), ModBlocks.witheredGrass.get().defaultBlockState(), Block.UPDATE_ALL);
+            } else {
+                this.setWilted(true);
+            }
+        }
     }
 
-    public boolean isWithered() {
-        return this.withered;
+    public void setWilted(boolean flag) {
+        this.wilted = flag;
+        if (this.level != null)
+            this.level.setBlock(this.getBlockPos(), this.level.getBlockState(this.getBlockPos()).setValue(BlockCrop.WILTED, flag), Block.UPDATE_ALL);
+    }
+
+    private void updateState() {
+        if (!(this.level instanceof ServerLevel))
+            return;
+        BlockState newState = this.getBlockState();
+        if (newState.getBlock() instanceof BlockCrop crop) {
+            float max = crop.properties().map(CropProperties::growth).orElse(0);
+            this.age = Math.min(crop.properties().map(CropProperties::growth).orElse(0), this.age);
+            int stateAge = (int) (this.age * 3 / max);
+            newState = newState.setValue(BlockCrop.AGE, Mth.clamp(stateAge, 0, 3))
+                    .setValue(BlockCrop.WILTED, this.wilted);
+        }
+        this.level.setBlock(this.getBlockPos(), newState, Block.UPDATE_ALL);
+    }
+
+    public boolean isWilted() {
+        return this.wilted;
     }
 
     @Override
@@ -95,7 +126,8 @@ public class CropBlockEntity extends BlockEntity {
         this.age = nbt.getFloat("Age");
         this.isGiant = nbt.getBoolean("Giant");
         this.cropLvl = nbt.getFloat("Level");
-        this.withered = nbt.getBoolean("Withered");
+        this.wilted = nbt.getBoolean("Wilted");
+        this.updateState();
     }
 
     @Override
@@ -104,6 +136,6 @@ public class CropBlockEntity extends BlockEntity {
         nbt.putFloat("Age", this.age);
         nbt.putBoolean("Giant", this.isGiant);
         nbt.putFloat("Level", this.cropLvl);
-        nbt.putBoolean("Withered", this.withered);
+        nbt.putBoolean("Wilted", this.wilted);
     }
 }