+++ fabric/src/main/java/io/github/flemmli97/runecraftory/fabric/config/ConfigHolder.java
@@ -43,6 +43,7 @@ public record ConfigHolder<T>(T configSpec, Consumer<T> loader) {
         GeneralConfig.disableFoodSystem = spec.disableFoodSystem.get();
         GeneralConfig.disableItemStatSystem = spec.disableItemStatSystem.get();
         GeneralConfig.disableCropSystem = spec.disableCropSystem.get();
+        GeneralConfig.witherChance = spec.witherChance.get().floatValue();
 
         GeneralConfig.maxLevel = spec.maxLevel.get();
         GeneralConfig.maxSkillLevel = spec.maxSkillLevel.get();